refactor: move HSQLDB, InMemoryLdapServer, SampleLocalMailbox to test sources. - add back native - put hsqldb as test dependency - default webapp /opt/zextras/conf
