+++ packages/yap.json
@@ -17,6 +17,9 @@
       "name": "common-appserver-conf"
     },
     {
+      "name": "common-appserver-native-lib"
+    },
+    {
       "name": "directory-server"
     },
     {
